refactor out VQClassification class w/ helper methods
