+++ qiskit/aqua/algorithms/adaptive/__init__.py
@@ -19,8 +19,12 @@ from .vqe.vqe import VQE
 from .qaoa.qaoa import QAOA
 from .qsvm.qsvm_variational import QSVMVariational
 from .vqalgorithm import VQAlgorithm
+from .vqclassification import VQClassification
 
-__all__ = ['VQE',
-           'QAOA',
-           'QSVMVariational',
-           'VQAlgorithm']
+__all__ = [
+    'VQE',
+    'QAOA',
+    'QSVMVariational',
+    'VQAlgorithm',
+    'VQClassification',
+]
