+++ qiskit/aqua/algorithms/adaptive/qsvm/__init__.py
@@ -14,9 +14,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # =============================================================================
-
-from .cost_helpers import cost_estimate, cost_estimate_sigmoid, return_probabilities
-
-__all__ = ['cost_estimate',
-           'cost_estimate_sigmoid',
-           'return_probabilities']
